+++ lib/MForm/Repeater/MFormRepeaterHelper.php
@@ -157,7 +157,7 @@ class MFormRepeaterHelper
     {
         if (empty($nameKey)) $nameKey = self::getNameKey($mformItem);
         $mformItem->addAttribute('x-model', $group . '[\'' . $nameKey . '\']')
-            ->addAttribute(':id', "'".$nameKey."-'+".$repeaterId."Index".((!empty($parentId))?"+'-'+".$parentId.'Index':''))
+            ->addAttribute(':id', "'".$nameKey.'-'.$repeaterId."-'+".$repeaterId."Index".((!empty($parentId))?"+'-".$parentId."-'+".$parentId.'Index':''))
             ->addAttribute('group', $group)
             ->addAttribute('groups', $groups)
             ->addAttribute('repeaterId', $repeaterId)