+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Repeater issues with addLinkField, addCustomLinkField, addMediaField, addInputs, addToggleCheckboxField, addCollapseElement, setToggleOptions solved
 - Repeater CustomLink yForm link option issue solved
 - Fix radio collapse not work in repeater issue 
+- cke5 reinit issue fixed
 
 ## Version 8.0.3
 - add showWrapper functionality